requirements.txt
#requirements.txt
crewai
requests
python-dotenv
markdown2
pyyaml
